Implement overall state monitor
Overal state monitor is an implementation of thread, functions and
structure that observe other modules and threads.
The idea of having this in the project is to provide automatic analyze
of some common problems and display to the user via different
interfaces. These interfaces can be LEDs, graphic display, serial
communication, 7-segment display or other UI solution.

This implementation is the first version of overall state monitor it can
be changed, improved or increased over the time.

Details:
 - Added thread and an enumerator member for the handler of this thread
 - Thread with simple checks detects if timebase and RTC clocks are
 working and sets the flag in the strucutre
 - After populating the structure thread calls function that analyzes
 strucutre data and in case of error or deviation reports in UI.

Issue-id: #34