blob-processor: do not increase counter when fail
After discovering a critical failure by a blob function, means that it
is better not to continue blob processing further. So, increasing the
counter has to be done only after passing critical check.

Critical check is blob->data.args_length value, it must be exactly that
long as we expect for this type of a function or determine differently
for irregular types of function: e.g. _print checks if the last argument
is '\0'.

If critical check fails, means that we don't know exactly when next
fucntion's blob_data starts and incrementing blob->counter can cause
unpredictable behaviour or blob processor failure.

Issue-id: #89