Fix cmdline task WM reaches 0 while setting date
When trying to set datetime via command line, cmdline thread watermark
raches 0. The reason is that datetime set uses `sscanf` function to parse
passed datetime arguments which is greedy when it comes to memory.

This patch increases stack depth for cmdline thread from 256 to 512 and
after `datetime set dd-MM-yy HH:mm:ss` unused bytes are 254, means before
this patch, stack of this thread was overflowed by two bytes.

Also, changed the name of the thread from `UART1 Rx` to `CMD line`, which
is more relevant to what the thread does.

Issue-id: #95