threads: fix adc thread takes all cpu resources
When ADC module was not initialized ADC thread was just switching
the context to return back for another check as soon as possible.

This patch suspends ADC thread if it discovers that ADC module is not
initialized. Suspended thread resumes from upper layer (in this
case command line interface) after initializing the module.

Issue-id: #124