wrappers: make UART module common for all UARTs
The old implementation of UART is removed. The new implementation is based on
UART configuration struct. This configuration can be modified during runtime
before the initialization of the peripheral device.
The usage of this new addition is visible in this patch in t_cmdline.c
and mikiot_config.h files.
This patch also contains implementations for the issue: #118

Issue-id: #136