ui: implement gprs command
The new command `gprs` can provide some useful functionality, like:
call, message, send data to the server, start server and run any command
with `gprs cmd` combination. It is possible to set timeout and an
expected data value as well.

The command has help in case of incorrectly (or not) providing
arguments.

Issue-id: #30