libs: ringbuffer: fix conditional and add tests
Fix: ringbuffer algorithm when it doesn't support overlap had incorrect
conditional. It has to block the thread until enqueue length is more
than available space.

Add: tests for overlap supporting mode of the ringbuffer. The mode when
overlap is not supported is harder to implement in the test, as it will
be blocking test indefinitely. Some features has to be implemented
before this test will be done, like: timer based delayed asynchronous
function execution or test with multiple threads.

P.S. Changed all tabs to spaces.

Issue-id: #139