cli: stop blocking when running the blob
The blob command is a blocking command. If blob has an infinite loop or
contains big delays it is impossible to use CLI. This patch provides the
solution, the command blob runs in separated thread and if it doesn't
finish, CLI is not blocked anymore.
Each instance of blob thread has its incremented unique name.

Closes #191