drivers: add PWM driver
This patch adds PWM driver and combines it with input sygnal reader
module (mgr_signal) in t_sigmgr. The changes are done in order to test
new PWM module with RC controller.

Due to difficulties with multiple interrupt lines in the same IRQ
prevents the precise measurement of signal length all interrupts are
removed from IO manager and assigned to signal manager module. IO
manager manages with TIMER interrupts now that was anyways running
already. Maybe it keeps the code in the interrupt slightly longer but
doesn't hurt anything so far.

With this setup it is possible to deal with up to 7 signals precisely.

Closes #186